+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 ## Changelog
 
+### v1.2.5
+
+- Add RSS item_pubdate
+- Address string formatting issue with tickets
+
+### v1.2.4
+
+- Add RSS feed for news/announcements
+- Add category field for tickets
+- Don't display the "untitled" attachments from RT, which is just the ticket/reply content
+- Add /robots.txt
+
+### v1.2.3
+
+- Add ability to request PI Eligibility from profile edit page
+- Added admin functionality
+
 ### v1.2.2
 
 - G5K Discovery fixes